Always set anonymous id, even if we have user id
Previously, the anonymous id was only set if we did not have a user id
(from current_user.id). However, the server requires the anonymous id
always be present as a UUID, so now it is generated/set as a cookie and
sent even if we have a user id.